requirements.txt
Flask==2.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
multipart==0.2.4
numpy==1.22.3
Pillow==9.0.1
redis==4.1.4
tensorflow-gpu==2.8.0

